// 2024/2024-05-clever-platform-backoffice/v1/src/context/domain/g05/g05-d01/g05-d01-p02-teacher-homework/local/type.ts
export enum Status {
  IN_USE = 'ใช้งาน',
  DRAFT = 'แบบร่าง',
  NOT_IN_USE = 'ไม่ใช้งาน',
}
// ...
export interface StudyGroup {
  study_group_id: number;
  study_group_name: string;
  selected?: boolean;
}

export interface ClassData {
  class_id: number;
  class_name: string;
  study_group: StudyGroup[];
  selected?: boolean;
}

export interface YearData {
  seed_year_id: number;
  seed_year_short_name: string;
  class: ClassData[];
  selected?: boolean;
}
// ...
export interface AssignedTo {
  class_ids: number[];
  study_group_ids: number[];
  // seed_year_ids: number[];
}

export interface CreateHomeworkRequest {
  name: string;
  // empty string. not affect any thing
  homework_template_name: '';
  homework_template_subject_id: number;
  homework_template_year_id: number;
  homework_template_lesson_id: number;
  homework_template_status: string;
  subject_id: number;
  year_id: number;
  started_at: string;
  due_at: string;
  closed_at: string;
  status: string;
  assigned_to: AssignedTo;
  level_ids?: number[];
}